backend/requirements.txt
fastapi
uvicorn
gunicorn
pytest
python-jose[cryptography]
passlib[bcrypt]
python-multipart